Split new_exec.h into new_exec.cc
